requirements_api.txt
# NFT Market Analytics API - Production Dependencies
# Pinned versions for reproducible deployments

# Web Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Configuration
python-dotenv==1.0.0

# Type Hints & Validation
pydantic==2.5.0
pydantic-core==2.14.1

# Optional: Monitoring & Logging
python-json-logger==2.0.7
